requirements.txt
attrs==23.2.0
beautifulsoup4==4.12.3
certifi==2024.2.2
cfgv==3.4.0
charset-normalizer==3.3.2
distlib==0.3.8
filelock==3.13.1
h11==0.14.0
identify==2.5.35
idna==3.6
isort==5.13.2
nodeenv==1.8.0
outcome==1.3.0.post0
packaging==23.2
platformdirs==4.2.0
pre-commit==3.6.2
pysocks==1.7.1
python-dotenv==1.0.1
pyyaml==6.0.1
requests==2.31.0
ruff==0.2.2
selenium==4.18.1
setuptools==69.1.1
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
trio==0.24.0
trio-websocket==0.11.1
typing-extensions==4.9.0
urllib3==2.2.1
virtualenv==20.25.1
webdriver-manager==4.0.1
wsproto==1.2.0
